fix: make vector add block-only PTX codegen works for float64 as well

+++ tests/test2.py
@@ -8,7 +8,7 @@ def kernel_prange(a, b):
         c[i] = a[i] + b[i]
     return c
 
-a = np.ones(4).astype(np.float32)
-b = np.ones(4).astype(np.float32)
+a = np.ones(4).astype(np.float64)
+b = np.ones(4).astype(np.float64)
 res = kernel_prange(a, b)  
 print(res)
